Publish npm pkg & error throw if boilerplate doesn't exists

+++ pengo-mango.js
@@ -2,17 +2,28 @@ const path = require("path");
 const util = require("util");
 const exec = util.promisify(require("child_process").exec);
 const chalk = require("chalk");
+const isFileExists = require("./src/isFileExists");
+
+const defaultBoilerplate = "penguin.js";
+const boilerplate = process.argv[2] || defaultBoilerplate;
 
 async function PengoMangoCLI() {
   try {
     const cwd = process.cwd();
-    const boilerplate = "penguin.js";
+    if (!(await isFileExists(boilerplate))) {
+      throw new Error(
+        `[ ] There has no template boilerplate exists named ${boilerplate}`
+      );
+    }
     const templatePath = path.resolve(`${__dirname}/templates/${boilerplate}`);
     const copyCommend = `cp -al ${templatePath}/. ${cwd}/.`;
     await exec(copyCommend);
-    console.log(chalk.white.bgGreen.bold(`Template Scaffolded Successfully`));
+    console.log(
+      chalk.white.bgGreen.bold(
+        `[x] ${boilerplate} Template Scaffolded Successfully`
+      )
+    );
   } catch (error) {
-    console.log(chalk.red(`Failed to start template`));
     console.log(chalk.red(error.message));
   }
 }
